+++ resources/handlers/workhour/get-one.ts
@@ -40,7 +40,7 @@ export async function getWorkHourFromUsernameBasedOnDate(
     startTimestamp: parseInt(item.startTimestamp.N!),
     endTimestamp: parseInt(item.endTimestamp.N!),
     date: item.sk.S!.split("#")[1], // Extract date from the sort key
-    duration: parseInt(item.duration.N!),
+    duration: parseFloat(item.duration.N!),
   };
 
   return {
